refactor : group-study 컴포넌트 도메인 폴더 통합
discussion, schedule, section, filtering, participation, premium, summary
폴더를 src/components/group-study/ 하위로 통합.
group-study 관련 pages, forms도 함께 이동하며 import 경로 일괄 수정.
