+++ src/components/common/modals/group-study-form-modal.tsx
@@ -11,7 +11,7 @@ import { useForm } from 'react-hook-form';
 
 import type { GroupStudyFullResponseDto } from '@/api/openapi';
 import { Modal } from '@/components/common/ui/modal';
-import GroupStudyForm from '@/components/forms/group-study-form';
+import GroupStudyForm from '@/components/group-study/forms/group-study-form';
 import { STUDY_TYPES, THUMBNAIL_EXTENSION } from '@/config/group-study-const';
 import { useAuthReady } from '@/features/auth/model/use-auth';
 import {
