+++ src/components/contents/payment-page-content.tsx
@@ -3,8 +3,8 @@
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import PaymentCheckoutPage from '@/components/payment/paymentActionClient';
-import OrderSummary from '@/components/summary/order-summary';
-import PriceSummary from '@/components/summary/price-summary';
+import OrderSummary from '@/components/group-study/summary/order-summary';
+import PriceSummary from '@/components/group-study/summary/price-summary';
 import { usePreparePaymentQuery } from '@/hooks/queries/payment/payment-user-api';
 import { useToastStore } from '@/stores/use-toast-store';
