+++ src/components/group-study/discussion/channel/post.tsx
