+++ src/components/group-study/filtering/study-search.tsx
