+++ src/components/group-study/forms/group-study-steps/group-study-step-introduction.tsx
