+++ src/components/group-study/forms/group-study-thumbnail-input.tsx
