+++ src/components/group-study/participation/reservation-user-card.tsx
