+++ src/components/group-study/schedule/today-study-card.tsx
