+++ src/components/group-study/section/curriculum-summary-section.tsx
