+++ src/components/group-study/section/inquiry-section.tsx
