+++ src/components/group-study/section/my-participating-studies-section.tsx
