+++ src/components/group-study/summary/study-info-summary.tsx
