+++ src/hooks/common/use-study-list-filter.ts
@@ -8,7 +8,7 @@ import type {
   GetGroupStudiesTargetRolesEnum,
   GetGroupStudiesTypeEnum,
 } from '@/api/openapi/api/group-study-management-api';
-import type { StudyFilterValues } from '@/components/filtering/study-filter';
+import type { StudyFilterValues } from '@/components/group-study/filtering/study-filter';
 import { useGetStudies } from '@/hooks/queries/one-to-one/study-query';
 
 const PAGE_SIZE = 15;
